refactor: 선언한 타입들 파일 분리
