+++ src/components/Ranking/RankItem.tsx
@@ -3,7 +3,7 @@ import { Flex, Text, Box } from '@chakra-ui/react';
 import DustState from '@/components/Dust/DustState';
 import { FINE_DUST, ROUTE, ULTRA_FINE_DUST } from '@/utils/constants';
 import { getDustAverageGrade } from '@/utils/dustGrade';
-import type { CityAirQuality } from '@/type';
+import type { CityAirQuality } from '@/types/dust';
 
 interface RankItemProps extends CityAirQuality {
   rank: number;
