--- src/type/index.ts
@@ -1,18 +1,0 @@
-export interface CityAirQuality {
-  cityName: string;
-  fineDustScale: number;
-  fineDustGrade: number;
-  ultraFineDustScale: number;
-  ultraFineDustGrade: number;
-  dataTime: string;
-}
-
-export interface Location {
-  latitude: number;
-  longitude: number;
-}
-
-export interface CityGroup {
-  cityName: string;
-  cityNumber: number;
-}
