+++ src/types/city.ts
@@ -1,0 +1,4 @@
+export interface CityGroup {
+  cityName: string;
+  cityNumber: number;
+}
