+++ src/types/dust.ts
@@ -7,3 +7,18 @@ export interface DustGrades {
   pm10Grade: string;
   pm25Grade: string;
 }
+
+export interface CityAirQuality {
+  cityName: string;
+  fineDustScale: number;
+  fineDustGrade: number;
+  ultraFineDustScale: number;
+  ultraFineDustGrade: number;
+  dataTime: string;
+}
+
+export type DustGradeType = 'GOOD' | 'NORMAL' | 'BAD' | 'DANGER';
+
+export interface DustGrade {
+  [key: number]: DustGradeType;
+}
