+++ src/types/location.ts
@@ -1,0 +1,4 @@
+export interface Location {
+  latitude: number;
+  longitude: number;
+}
