+++ src/utils/constants/dust.ts
@@ -1,13 +1,9 @@
+import type { DustGrade } from '@/types/dust';
+
 export const FINE_DUST = '미세먼지';
 export const ULTRA_FINE_DUST = '초미세먼지';
 
-export type GradeType = 'GOOD' | 'NORMAL' | 'BAD' | 'DANGER';
-
-export interface Grade {
-  [key: number]: GradeType;
-}
-
-export const DUST_GRADE: Grade = {
+export const DUST_GRADE: DustGrade = {
   1: 'GOOD',
   2: 'NORMAL',
   3: 'BAD',
